move app.ts to the root
