+++ src/app.ts
@@ -5,17 +5,17 @@ import { createHttpTerminator, type HttpTerminator } from "http-terminator";
 import yaml from "js-yaml";
 import $RefParser from "json-schema-ref-parser";
 
-import { startRepl } from "../repl/repl.js";
-import { CodeGenerator } from "../typescript-generator/code-generator.js";
-import { readFile } from "../util/read-file.js";
-import type { Config } from "./config.js";
-import { ContextRegistry } from "./context-registry.js";
-import { createKoaApp } from "./create-koa-app.js";
-import { Dispatcher, type OpenApiDocument } from "./dispatcher.js";
-import { koaMiddleware } from "./koa-middleware.js";
-import { ModuleLoader } from "./module-loader.js";
-import { Registry } from "./registry.js";
-import { Transpiler } from "./transpiler.js";
+import { startRepl } from "./repl/repl.js";
+import type { Config } from "./server/config.js";
+import { ContextRegistry } from "./server/context-registry.js";
+import { createKoaApp } from "./server/create-koa-app.js";
+import { Dispatcher, type OpenApiDocument } from "./server/dispatcher.js";
+import { koaMiddleware } from "./server/koa-middleware.js";
+import { ModuleLoader } from "./server/module-loader.js";
+import { Registry } from "./server/registry.js";
+import { Transpiler } from "./server/transpiler.js";
+import { CodeGenerator } from "./typescript-generator/code-generator.js";
+import { readFile } from "./util/read-file.js";
 
 async function loadOpenApiDocument(source: string) {
   try {
